Move search expressions utils to NP!
