+++ src/legacy/core_plugins/kibana/public/discover/kibana_services.ts
@@ -50,16 +50,15 @@ export function setServices(newServices: any) {
 // EXPORT legacy static dependencies, should be migrated when available in a new version;
 export { angular };
 export { wrapInI18nContext } from 'ui/i18n';
+import { search } from '../../../../../plugins/data/public';
+export const { getRequestInspectorStats, getResponseInspectorStats, tabifyAggResponse } = search;
 // @ts-ignore
 export { shortenDottedString } from '../../common/utils/shorten_dotted_string';
-export { getRequestInspectorStats, getResponseInspectorStats } from '../../../data/public';
 // @ts-ignore
 export { intervalOptions } from 'ui/agg_types';
 export { subscribeWithScope } from 'ui/utils/subscribe_with_scope';
 // @ts-ignore
 export { timezoneProvider } from 'ui/vis/lib/timezone';
-import { search } from '../../../../../plugins/data/public';
-export const { tabifyAggResponse } = search;
 export { unhashUrl } from '../../../../../plugins/kibana_utils/public';
 export {
   ensureDefaultIndexPattern,
