+++ src/plugins/expressions/public/index.ts
@@ -94,6 +94,7 @@ export {
   KibanaContext,
   KibanaDatatable,
   KibanaDatatableColumn,
+  KibanaDatatableColumnMeta,
   KibanaDatatableRow,
   KnownTypeToString,
   Overflow,
